Post_services2/Backend/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
SQLAlchemy==2.0.42
typing_extensions==4.14.1
Werkzeug==3.1.3
flask-cors==6.0.1
PyJWT>=2.9.0
